# frontend/openapispec.yml
openapi: 3.0.1
info:
  title: leagueoflegeds
  version: '1.0'
servers:
  - url: 'http://localhost:4000'
paths:
  /auth/register:
    post:
      summary: ''
      operationId: post-auth-register
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                password:
                  type: string
                summonerName:
                  type: string
                region:
                  type: string
              required:
                - userName
                - password
                - summonerName
                - region
      description: a Endpiont to register a new user
  /auth/login:
    post:
      summary: ''
      operationId: post-auth-login
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                password:
                  type: string
              required:
                - userName
                - password
      description: 'a Endpiont to login with userName and password '
  /auth/update:
    put:
      summary: ''
      operationId: put-auth-update
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      parameters:
        - schema:
            type: string
          in: query
          name: region
        - schema:
            type: string
          in: query
          name: summonerName
        - schema:
            type: string
          in: query
          name: userName
      description: update summonerName for user
...
components:
  schemas:
    User:
      title: User
      type: object
      description: ''
      properties:
        id:
          type: string
          description: Unique identifier for the given user.
          format: uuid
        userName:
          type: string
        followers:
          type: array
          items:
            $ref: '#/components/schemas/Follower'
        summonerName:
          $ref: '#/components/schemas/SummonerWithQueuesMatches'
      required:
        - id
        - userName
        - summonerName
    Summoner:
      title: Summoner
      allOf:
        - properties:
            id:
              type: string
            puuid:
              type: string
            name:
              type: string
            region:
              type: string
            profileIconId:
              type: integer
            summonerLevel:
              type: number
          required:
            - id
            - puuid
            - name
            - region
            - profileIconId
            - summonerLevel
      type: object
    Follower:
      title: Follower
      allOf:
        - properties:
            id:
              type: string
            followerName:
              type: string
            region:
              type: string
            profileIconId:
              type: integer
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'
      type: object
    Queue:
      title: Queue
      allOf:
        - properties:
            queueType:
              type: string
            tier:
              type: string
            rank:
              type: string
            leaguePoints:
              type: integer
            wins:
              type: integer
            losses:
              type: integer
          required:
            - queueType
            - tier
            - rank
            - leaguePoints
            - wins
            - losses
      type: object
    Match:
      title: Match
      allOf:
        - properties:
            id:
              type: string
            evaluations:
              type: array
              items:
                $ref: '#/components/schemas/Evaluation'
          required:
            - id
      type: object
    SummonerWithQueuesMatches:
      title: SummonerWithQueueMatch
      allOf:
        - $ref: '#/components/schemas/Summoner'
        - type: object
          properties:
            queues:
              type: array
              items:
                $ref: '#/components/schemas/Queue'
            matches:
              type: array
              items:
                $ref: '#/components/schemas/Match'
...
    Evaluation:
      title: Evaluation
      type: object
      description: Evaluation for a Match
      properties:
        summonerName:
          type: string
        tag:
          type: string
        description:
          type: string
        feedback:
          type: string
        lane:
          type: string
...